+++ client/test/e2e/tree-visibility.spec.ts
@@ -1,4 +1,3 @@
-// @ts-nocheck
 import { test, expect } from '@playwright/test';
 
 test('tree page shows niche carousel cards', async ({ page }) => {
